backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
python-magic==0.4.27
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2